bench: Add crc32 benchmark files for instruction counting.
Add three simple programs that operate on variable length input memory
files to be used in conjunction with the instruction count capability
to assist in the benchmarking of the eBPF VM and/or JITed code.

In all cases you should compile these C programs with clang and target
ebpf as the output machine. You can then run these programs using the
test executable and add the -i flag to perform instruction counting.