ubpf_inst_cnt: Add instruction counting.
Add a new struct that can carry information about the various types of
instruction counting we wish to perform.

In order to count the instructions executed on the underlying
bare-metal machine we leverage the perf_event framework (Linux
only). This uses ioctls to initialize and then measure the relevant
counters to ascertain how many instrucitons have been executing on the
CPU.

Note that access to these counters requires elevated privileges so
that may limit who can run the program in this mode.

Fixes #83.