Fix ubpf_plugin passing ambiguous pointer to memory for ubpf_exec
When there is no user-specified starting contents for the memory when
using ubpf_plugin, the vector representing the contents of the memory is
empty. There was an assumption that when the size of a std::vector is 0
then std::vector::data will return a null pointer but this is not
guaranteed to be the case.

Issue found while trying to debug tests/mul-loop.data failure in wasm
platform.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>