Rationalize parameters to helper functions
The external helper functions should get 6 parameters no matter whether
the eBPF program is interpeted or JIT'd. The 6 parameter, the one in
addition to the 5 specified by the eBPF standard, is the context
parameter.

Add tests to make sure that the 6th parameter is properly passed to the
external helper functions.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>