Fix stack usage validation for eBPF local functions
Stack usage for eBPF local functions was incorrectly calculated whether
or not a custom stack usage calculator was present or not. This PR fixes
that issue.

In addition to fixing that bug, this PR also changes the default eBPF
program stack size and the default stack size for local functions in
eBPF programs (8K and 256 bytes, respectively).

This PR also adds support for a stack usage calculator that does not
want to allow an eBPF program's local functions to use any stack space
at all.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>